chore: moving up main.py to the root directory + fixing some lints
